Patrick/results

+++ openroad_interface/openroad_run.py
@@ -21,7 +21,7 @@ from . import scale_lef_files as scale_lef
 DIE_CORE_BUFFER_SIZE = 50
 
 class OpenRoadRun:
-    def __init__(self, cfg, codesign_root_dir):
+    def __init__(self, cfg, codesign_root_dir, tmp_dir, run_openroad):
         """
         Initialize the OpenRoadRun with configuration and root directory.
 
@@ -30,7 +30,9 @@ class OpenRoadRun:
         """
         self.cfg = cfg
         self.codesign_root_dir = codesign_root_dir
-        self.directory = os.path.join(self.codesign_root_dir, "src/tmp/pd")
+        self.tmp_dir = tmp_dir
+        self.directory = os.path.join(self.codesign_root_dir, f"{self.tmp_dir}/pd")
+        self.run_openroad = run_openroad
 
     def run(
     self,
@@ -120,22 +122,25 @@ class OpenRoadRun:
         """
 
         logger.info("Setting up environment for place and route.")
-        if os.path.exists(self.directory):
-            logger.info(f"Removing existing directory: {self.directory}")
-            shutil.rmtree(self.directory)
-        os.makedirs(self.directory)
-        logger.info(f"Created directory: {self.directory}")
-        shutil.copytree(os.path.dirname(os.path.abspath(__file__)) + "/tcl", self.directory + "/tcl")
-        logger.info(f"Copied tcl files to {self.directory}/tcl")
-        os.makedirs(self.directory + "/results")
-        logger.info(f"Created results directory: {self.directory}/results")
+        if self.run_openroad:
+            if os.path.exists(self.directory):
+                logger.info(f"Removing existing directory: {self.directory}")
+                shutil.rmtree(self.directory)
+            os.makedirs(self.directory)
+            logger.info(f"Created directory: {self.directory}")
+            shutil.copytree(os.path.dirname(os.path.abspath(__file__)) + "/tcl", self.directory + "/tcl")
+            logger.info(f"Copied tcl files to {self.directory}/tcl")
+            os.makedirs(self.directory + "/results")
+            logger.info(f"Created results directory: {self.directory}/results")
+        else:
+            logger.info("Skipping setup, using previous openroad results.")
 
         self.update_area_constraint(area_constraint)
 
-        self.do_scale_lef = scale_lef.ScaleLefFiles(self.cfg, self.codesign_root_dir)
+        self.do_scale_lef = scale_lef.ScaleLefFiles(self.cfg, self.codesign_root_dir, self.tmp_dir)
         self.do_scale_lef.scale_lef_files(L_eff)
 
-        df = def_generator.DefGenerator(self.cfg, self.codesign_root_dir, self.do_scale_lef.NEW_database_units_per_micron)
+        df = def_generator.DefGenerator(self.cfg, self.codesign_root_dir, self.tmp_dir, self.do_scale_lef.NEW_database_units_per_micron)
 
         graph, net_out_dict, node_output, lef_data, node_to_num, area_estimate = df.run_def_generator(
             test_file, graph
@@ -358,7 +363,10 @@ class OpenRoadRun:
 
         # run openroad
         logger.info("Starting estimated place and route.")
-        self.run_openroad_executable()
+        if self.run_openroad:
+            self.run_openroad_executable()
+        else:
+            logger.info("Skipping openroad run.")
 
         wire_length_df = est.parse_route_guide_with_layer_breakdown(self.directory + "/results/codesign_codesign-tcl.route_guide")
         wire_length_by_edge = {}
@@ -387,10 +395,13 @@ class OpenRoadRun:
                     src = graph.nodes[node_name]["name"]
                     dst = graph.nodes[output_name]["name"]
                     #logger.info(f"Src: {src}, Dst: {dst}")
-                    if (src, dst) not in wire_length_by_edge:
-                        wire_length_by_edge[(src, dst)] = copy.deepcopy(wire_length_df.loc[net_name])
-                    else:
-                        wire_length_by_edge[(src, dst)] += copy.deepcopy(wire_length_df.loc[net_name])
+                    if net_name in wire_length_df.index:
+                        if (src, dst) not in wire_length_by_edge:
+                            wire_length_by_edge[(src, dst)] = copy.deepcopy(wire_length_df.loc[net_name])
+                        else:
+                            wire_length_by_edge[(src, dst)] += copy.deepcopy(wire_length_df.loc[net_name])
+                    #else:
+                    #    logger.warning(f"Net {net_name} not found in wire length dataframe. Skipping.")
         self.export_graph(graph, "estimated_with_mux")
 
         wire_length_by_edge = self.mux_listing(graph, node_output, wire_length_by_edge)
@@ -443,7 +454,10 @@ class OpenRoadRun:
 
         # run openroad
         logger.info("Starting detailed place and route.")
-        self.run_openroad_executable()
+        if self.run_openroad:
+            self.run_openroad_executable()
+        else:
+            logger.info("Skipping openroad run, as resource constraints have been reached in a previous iteration.")
 
         # run parasitic_calc and length_calculations
         graph, _ = self.coord_scraping(graph, node_to_num)
